FDN-1976: Install aws-cli from Alpine rather than pip
Installing the AWS cli via pip [no longer works][build] as pip
refuses to install packages globally because that would clobber the
distribution-managed Python packages.

It's possible to install via pip using a virtualenv but then one
has to activate it prior to use, so we probably would have to wrap
invocations of the aws command to activate the virtual env first.
I'd prefer to install via apk instead. The version instealled by apk
is 2.13.25-r0 while the version installed via pip is 1.32.11 and
there are [breaking changes][changes] so there's some risk.

[build]: https://jenkins.flo.pub/blue/organizations/jenkins/flowcommerce%2Fdocker/detail/main/34/pipeline/
[changes]: https://docs.aws.amazon.com/cli/latest/userguide/cliv2-migration-changes.html#cliv2-migration-changes-breaking