FDN-3609: do not build the "tools" repository in builder images
This will allow to upgrade the Go compiler and dependencies in the tools repository.
Otherwise we're stuck with Go 1.18 from the base image (Ubuntu jammy).

### Testing

I figured if the `dev` binary is used in some builder image, it would be in some Jenkinsfile.
So I did a search for that, with turned up empty:

	csearch -f Jenkinsfile 'dev '

The search string contains a space to avoid noise.
It's plausible, as usages of `dev` would be followed by a subcommand, i.e., `dev env`.